// frontend/types/api/models/Address.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Address = {
    readonly id: number;
    street: string;
    city: string;
    state: string;
    zip_code: string;
    readonly formatted_address: string;
};